+++ src/main.ts
@@ -6,8 +6,8 @@ import {
 
 export { AuthModule } from './auth/auth.module';
 export { AuthService } from './auth/auth.service';
-export { JwtAuthGuard } from './auth/jwt-auth.guard';
-export { LocalAuthGuard } from './auth/local-auth.guard';
+export { JwtAuthGuard } from './auth/jwt/jwt-auth.guard';
+export { LocalAuthGuard } from './auth/local/local-auth.guard';
 
 export async function bootstrap(version, appModule) {
   const app = await NestFactory.create<NestFastifyApplication>(
